Move everything to package specific to this plugin
