+++ src/main/java/me/element/customdeathmessages/enums/VersionEnums.java
@@ -1,4 +1,4 @@
-package me.element.enums;
+package me.element.customdeathmessages.enums;
 
 public enum VersionEnums {
