+++ src/main/java/me/element/customdeathmessages/other/CustomConsumer.java
@@ -1,4 +1,4 @@
-package me.element.other;
+package me.element.customdeathmessages.other;
 
 public interface CustomConsumer<T> {
   void accept(T paramT);
