+++ src/main/java/me/element/customdeathmessages/other/JsonChat.java
@@ -1,4 +1,4 @@
-package me.element.other;
+package me.element.customdeathmessages.other;
 
 import java.lang.reflect.Method;
